CORE-10592 Renamed 'permission-rpc-ops-impl' to 'permission-rest-resource-impl (#3201)
