+++ components/permissions/permission-rest-resource-impl/build.gradle
@@ -1,0 +1,35 @@
+plugins {
+    id 'corda.common-publishing'
+    id 'corda.common-library'
+}
+
+description "Permission REST Resource Implementations"
+
+dependencies {
+    compileOnly 'org.osgi:osgi.annotation'
+    compileOnly "org.osgi:org.osgi.service.component.annotations"
+
+    implementation platform("net.corda:corda-api:$cordaApiVersion")
+    implementation "net.corda:corda-base"
+    implementation "net.corda:corda-avro-schema"
+    implementation "net.corda:corda-topic-schema"
+
+    implementation project(":libs:configuration:configuration-core")
+    implementation project(':libs:messaging:messaging')
+    implementation project(':libs:permissions:permission-endpoint')
+    implementation project(':libs:permissions:permission-manager')
+    implementation project(':libs:permissions:permission-common')
+    implementation project(":libs:rest:rest-common")
+    implementation project(":libs:rest:rest")
+    implementation project(":libs:utilities")
+
+    implementation project(':components:permissions:permission-management-service')
+
+    implementation project(":libs:lifecycle:lifecycle")
+
+    implementation 'org.jetbrains.kotlin:kotlin-osgi-bundle'
+    implementation "org.slf4j:slf4j-api:$slf4jVersion"
+
+    testImplementation "org.mockito:mockito-inline:$mockitoInlineVersion"
+    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
+}
