+++ components/permissions/permission-rest-resource-impl/src/main/kotlin/net/corda/libs/permissions/endpoints/common/PermissionEndpointEventHandler.kt
