+++ components/permissions/permission-rest-resource-impl/src/test/kotlin/net/corda/libs/permissions/endpoints/common/PermissionEndpointEventHandlerTest.kt
