+++ settings.gradle
@@ -158,7 +158,7 @@ include 'components:permissions:permission-storage-reader-service'
 include 'components:permissions:permission-storage-writer-service'
 include 'components:permissions:permission-validation-service'
 include 'components:permissions:permission-validation-cache-service'
-include 'components:permissions:permission-rpc-ops-impl'
+include 'components:permissions:permission-rest-resource-impl'
 include 'components:persistence:entity-processor-service'
 include 'components:persistence:entity-processor-service-impl'
 include 'components:persistence:persistence-service-common'
